fix(ci): python full coverage — oversized provisioner test shard exceeded line limit
